Allow the region tooltip to be shown with the PlaceOrders popover as well
